Diffusion Coefficient for liquid molecules only
This python3 code calculates the diffusion coefficient from a simulation of only liquid molecules. The liquid molecules were modelled as hard spheres to save computing time and to make the problem less complex. The diffusion coefficient equation id D = <MSD>/2nt, where D is the diffusion coefficient, MSD is the mean squared displacement, the <> brackets mean an average is taken, n is the number of dimensions the liquid can travel in in this case 3 dimensions, t is the time period of the simulation. A graph of <MSD> (Angstroms) against time (picoseconds) is calculated. The gradient of said graph would equal to 2Dn. Using this the Diffusion coefficient can be calculated.