Solid Wall
This code places the molecules into a face centred cubic arrangement for one of the simulation input files. This is to keep the energy of the system low. The code creates 4 layers currently however this is easily changed. The solid layer starts in the middle of the simulation cell. This to provide better averaging in the solid-liquid simulation.